requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMuPDF==1.23.8
google-generativeai==0.3.2
Werkzeug==2.3.7
gunicorn==21.2.0
